Revert unintended route and action changes
